Factor out requires valid user url matcher test
